computeに圧縮解凍処理を統合
